# domains/core-platform-services/auth-service/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/academy/enrollment": {
      "post": {
        "description": "👤 Student enrollment into a course or cohort",
        "operationId": "EnrollmentController_createEnrollment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEnrollmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Enrollment created successfully"
          },
          "400": {
            "description": "Invalid enrollment data"
          }
        },
        "summary": "Enroll in a course",
        "tags": [
          "Enrollments"
        ]
      },
      "get": {
        "description": "🔒 Admin / Instructor",
        "operationId": "EnrollmentController_findAllEnrollments",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all enrollments",
        "tags": [
          "Enrollments"
        ]
      }
    },
...
    "/academy/enrollment/cohort/{cohortId}": {
      "get": {
        "description": "🔒 Instructor only",
        "operationId": "EnrollmentController_findEnrollmentsByCohort",
        "parameters": [
          {
            "name": "cohortId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get enrollments by cohort",
        "tags": [
          "Enrollments"
        ]
      }
    },
    "/academy/enrollment/{id}": {
      "delete": {
        "description": "🔒 Admin only",
        "operationId": "EnrollmentController_removeEnrollment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Remove enrollment",
        "tags": [
          "Enrollments"
        ]
      }
    },
    "/academy/enrollment/user/{userId}": {
      "get": {
        "description": "🔒 Admin only",
        "operationId": "EnrollmentController_findEnrollmentsByUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get enrollments by user ID",
        "tags": [
          "Enrollments"
        ]
      }
    },
...
    "/academy/progress/course/{courseId}/lesson/{lessonId}/complete": {
      "post": {
        "operationId": "ProgressAndAnalyticsController_completeLesson",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Mark lesson as complete for the current user",
        "tags": [
          "Progress & Analytics"
        ]
      }
    },
    "/academy/progress/course/{courseId}/analytics": {
      "get": {
        "description": "🔒 Instructor / Admin only",
        "operationId": "ProgressAndAnalyticsController_courseAnalytics",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get course analytics",
        "tags": [
          "Progress & Analytics"
        ]
      }
    },
    "/academy/progress/course/{courseId}/students": {
      "get": {
        "description": "🔒 Instructor / Admin only",
        "operationId": "ProgressAndAnalyticsController_studentsProgress",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all students progress for a course",
        "tags": [
          "Progress & Analytics"
        ]
      }
    },
...
    "/academy/teaching-schedules/course/{courseId}": {
      "get": {
        "description": "Accessible by enrolled users, instructors, or admin",
        "operationId": "TeachingScheduleController_getCourseSchedules",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all schedules for a specific course",
        "tags": [
          "Teaching Schedules"
        ]
      }
    },
    "/academy/teaching-schedules": {
      "post": {
        "description": "🔒 Instructor / Admin only",
        "operationId": "TeachingScheduleController_createSchedule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeachingScheduleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a teaching schedule",
        "tags": [
          "Teaching Schedules"
        ]
      }
    },
    "/academy/teaching-schedules/{id}": {
      "put": {
        "description": "🔒 Instructor / Admin only",
        "operationId": "TeachingScheduleController_updateSchedule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeachingScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a teaching schedule",
        "tags": [
          "Teaching Schedules"
        ]
      },
      "delete": {
        "description": "🔒 Instructor / Admin only",
        "operationId": "TeachingScheduleController_deleteSchedule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Delete a teaching schedule",
        "tags": [
          "Teaching Schedules"
        ]
      },
      "get": {
        "description": "Accessible by instructor/admin or enrolled users",
        "operationId": "TeachingScheduleController_getScheduleById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a teaching schedule by ID",
        "tags": [
          "Teaching Schedules"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Gateway",
    "description": "Gateway documentation for frontend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateEnrollmentDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "User ID (optional, derived from token if omitted)",
            "example": "firebase_uid_123"
          },
          "cohortId": {
            "type": "number",
            "description": "Cohort ID (optional)",
            "example": 5
          },
          "courseId": {
            "type": "number",
            "description": "Course ID",
            "example": 10
          }
        },
        "required": [
          "courseId"
        ]
      },
...
      "CreateTeachingScheduleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the schedule",
            "example": "NestJS Live Class"
          },
          "description": {
            "type": "string",
            "description": "Optional description of the schedule",
            "example": "Deep dive into NestJS modules"
          },
          "startTime": {
            "type": "string",
            "description": "Start time of the schedule",
            "example": "2025-12-20T10:00:00Z"
          },
          "endTime": {
            "type": "string",
            "description": "End time of the schedule",
            "example": "2025-12-20T11:30:00Z"
          },
          "type": {
            "type": "string",
            "description": "Type of schedule",
            "enum": [
              "LIVE",
              "RECORDING",
              "Q_AND_A",
              "OFFICE_HOURS",
              "WORKSHOP"
            ],
            "example": "LIVE"
          },
          "courseId": {
            "type": "number",
            "description": "Associated course ID",
            "example": 1
          },
          "isRecurring": {
            "type": "boolean",
            "description": "Whether the schedule is recurring",
            "example": false
          },
          "recurrencePattern": {
            "type": "string",
            "description": "Recurrence pattern (if recurring)",
            "example": "WEEKLY"
          }
        },
        "required": [
          "title",
          "startTime",
          "endTime",
          "type",
          "courseId"
        ]
      },
      "UpdateTeachingScheduleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Updated title",
            "example": "Updated Live Class"
          },
          "description": {
            "type": "string",
            "description": "Updated description",
            "example": "Updated description content"
          },
          "startTime": {
            "type": "string",
            "description": "Updated start time",
            "example": "2025-12-21T10:00:00Z"
          },
          "endTime": {
            "type": "string",
            "description": "Updated end time",
            "example": "2025-12-21T11:30:00Z"
          },
          "type": {
            "type": "string",
            "description": "Updated type of schedule",
            "enum": [
              "LIVE",
              "RECORDING",
              "Q_AND_A",
              "OFFICE_HOURS",
              "WORKSHOP"
            ]
          },
          "isRecurring": {
            "type": "boolean",
            "description": "Updated recurring status",
            "example": true
          },
          "recurrencePattern": {
            "type": "string",
            "description": "Updated recurrence pattern",
            "example": "DAILY"
          }
        }
      },
...
    }
  }
}